Refactor activity creation workflow
