+++ client/src/lib/activities/__tests__/createActivity.test.ts
@@ -1,5 +1,26 @@
+import { jest } from "@jest/globals";
+
 import { ATTENDEE_REQUIRED_MESSAGE, END_TIME_AFTER_START_MESSAGE } from "@shared/activityValidation";
 
+jest.mock("@/lib/queryClient", () => ({
+  ApiError: class MockApiError extends Error {
+    status: number;
+    data: unknown;
+
+    constructor(status: number, data: unknown, message: string) {
+      super(message);
+      this.status = status;
+      this.data = data;
+    }
+  },
+  apiRequest: jest.fn(),
+}));
+
+import {
+  ActivitySubmissionError,
+  prepareActivitySubmission,
+  mapApiErrorToValidation,
+} from "../activityCreation";
 import { mapClientErrorToValidation } from "../clientValidation";
 
 describe("mapClientErrorToValidation", () => {
@@ -26,5 +47,68 @@ describe("mapClientErrorToValidation", () => {
     const fallback = mapClientErrorToValidation(null);
     expect(fallback.fieldErrors).toEqual([]);
     expect(fallback.formMessage).toMatch(/We couldn’t prepare your activity/);
+  });
+});
+
+describe("prepareActivitySubmission", () => {
+  const baseValues = {
+    name: "Morning Hike",
+    description: "Enjoy the sunrise",
+    startDate: "2024-05-01",
+    startTime: "08:00",
+    endTime: "10:00",
+    location: "Trailhead",
+    cost: "25",
+    maxCapacity: "8",
+    attendeeIds: ["1", "2"],
+    category: "outdoor",
+    type: "SCHEDULED" as const,
+  };
+
+  it("builds a submission payload for valid values", () => {
+    const { payload, sanitizedValues } = prepareActivitySubmission({
+      tripId: 42,
+      values: baseValues,
+    });
+
+    expect(payload).toMatchObject({
+      tripCalendarId: 42,
+      name: "Morning Hike",
+      mode: "scheduled",
+      attendeeIds: ["1", "2"],
+    });
+    expect(sanitizedValues.attendeeIds).toEqual(["1", "2"]);
+    expect(sanitizedValues.endTime).toBe("10:00");
+  });
+
+  it("throws an ActivitySubmissionError when validation fails", () => {
+    expect(() =>
+      prepareActivitySubmission({
+        tripId: 42,
+        values: { ...baseValues, startTime: "" },
+      }),
+    ).toThrow(ActivitySubmissionError);
+  });
+});
+
+describe("mapApiErrorToValidation", () => {
+  it("maps API validation errors to client fields", () => {
+    const apiError = {
+      status: 400,
+      data: { errors: [{ field: "start_time", message: "Start required" }] },
+    } as Parameters<typeof mapApiErrorToValidation>[0];
+
+    const validation = mapApiErrorToValidation(apiError);
+    expect(validation).toEqual({
+      fieldErrors: [{ field: "startTime", message: "Start required" }],
+      formMessage: undefined,
+    });
+  });
+
+  it("returns null for non-validation errors", () => {
+    const apiError = { status: 500, data: { message: "Server error" } } as Parameters<
+      typeof mapApiErrorToValidation
+    >[0];
+    expect(mapApiErrorToValidation(apiError)).toBeNull();
   });
 });
